Renamed dataFile to datafile package. Move datafile to the application layer.
