+++ abixen-platform-business-intelligence-service/src/main/java/com/abixen/platform/service/businessintelligence/multivisualisation/application/dto/datafile/DataFileDto.java
@@ -1,4 +1,3 @@
-
 /**
  * Copyright (c) 2010-present Abixen Systems. All rights reserved.
  *
@@ -12,25 +11,35 @@
  * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
  * details.
  */
-package com.abixen.platform.service.businessintelligence.multivisualisation.domain.model.util.dataFile;
+
+package com.abixen.platform.service.businessintelligence.multivisualisation.application.dto.datafile;
 
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class RowDto {
+public class DataFileDto {
 
-    private List<ColumnDto> columns = new ArrayList<>();
+    private RowDto rowTypes;
+    private List<RowDto> rows = new ArrayList<>();
 
-    public List<ColumnDto> getColumns() {
-        return columns;
+    public RowDto getRowTypes() {
+        return rowTypes;
     }
 
-    public void setColumns(List<ColumnDto> columns) {
-        this.columns = columns;
+    public void setRowTypes(RowDto rowTypes) {
+        this.rowTypes = rowTypes;
     }
 
-    public void addColumn(ColumnDto column) {
-        columns.add(column);
+    public List<RowDto> getRows() {
+        return rows;
+    }
+
+    public void setRows(List<RowDto> rows) {
+        this.rows = rows;
+    }
+
+    public void addRow(RowDto rowDto) {
+        rows.add(rowDto);
     }
 }
